Feature/#77: Core Module 추가
