+++ Harubee-iOS/Sources/Presentation/Today/TodayView.swift
@@ -7,6 +7,7 @@
 //
 
 import SwiftUI
+import Core
 import DesignSystem
 
 // MARK: - TodayView
